backend/requirements.txt
fastapi == 0.120.2
uvicorn == 0.38.0
python-jose == 3.5.0
requests == 2.32.5
httpx == 0.28.1
python-dotenv == 1.2.1
SQLAlchemy == 2.0.44
alembic == 1.17.0
psycopg2-binary == 2.9.11